sewing buff very magnificent and beautiful

+++ code/game/objects/items/rogueitems/needle.dm
@@ -36,6 +36,8 @@
 	..()
 
 /obj/item/needle/use(used)
+	if(infinite)
+		return TRUE
 	stringamt = stringamt - used
 	if(stringamt <= 0)
 		qdel(src)
